+++ packages/camera/camera/lib/src/camera_controller.dart
@@ -576,6 +576,7 @@ class CameraController extends ValueNotifier<CameraValue> {
   Future<void> startVideoRecording({
     onLatestImageAvailable? onAvailable,
     bool enablePersistentRecording = true,
+    String? videoOutputPath,
   }) async {
     _throwIfNotInitialized('startVideoRecording');
     if (value.isRecordingVideo) {
@@ -598,6 +599,7 @@ class CameraController extends ValueNotifier<CameraValue> {
           _cameraId,
           streamCallback: streamCallback,
           enablePersistentRecording: enablePersistentRecording,
+          videoOutputPath: videoOutputPath,
         ),
       );
       value = value.copyWith(