+++ packages/camera/camera/test/camera_preview_test.dart
@@ -113,6 +113,7 @@ class FakeController extends ValueNotifier<CameraValue>
   Future<void> startVideoRecording({
     onLatestImageAvailable? onAvailable,
     bool enablePersistentRecording = true,
+    String? videoOutputPath,
   }) async {}
 
   @override