+++ packages/camera/camera_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.14.0
+
+* Adds `videoOutputPath` support to `MethodChannelCamera.startVideoRecording`.
+
 ## 2.13.0
 
 * Updates minimum supported SDK version to Flutter 3.35/Dart 3.9.